added full screen toggle on f11
